+++ src/shared/consts/misc.js
@@ -16,3 +16,5 @@ export const legacyMode = settings.legacyMode === false ? false : true;
 
 // https://github.com/sindresorhus/semver-regex
 export const semverRegex = /(?<=^v?|\sv?)(?:0|[1-9]\d*)\.(?:0|[1-9]\d*)\.(?:0|[1-9]\d*)(?:-(?:[1-9]\d*|[\da-z-]*[a-z-][\da-z-]*)(?:\.(?:[1-9]\d*|[\da-z-]*[a-z-][\da-z-]*))*)?(?:\+[\da-z-]+(?:\.[\da-z-]+)*)?(?=$|\s)/gi;
+
+export const userIdRegex = /^[\da-f]+$/;